OCP-1160: delete redundant hook
